chore(core): Clean up dependencies (#324)
* chore(google): Clean fiat-google-groups dependencies

Use the spinnaker-dependencies provided google-api-client. Also
add in bootWeb and commonsLang that are directly used but were
only included transitively.  Finally, remove a number of compileOnly
dependencies that are not used.

* chore(github): Clean fiat-github dependencies

Add some required dependencies that were only included transitively,
and remove a few dependencies that were not used.

* chore(roles): Clean fiat-roles dependencies

Push google-api-client to spinnaker-dependencies, and remove
two unused dependencies.

* chore(roles): Remove durian dependency

Was looking into whether to bump the durian dependency, but it's
only used in a single place and the same functionality exists
directly in Java 8, so just removed the dependency.

* fix(core): Remove duplicate dependency